Move core datastructures to 'core' module
Additionally split out views from anndata.py
