+++ anndata/__init__.py
@@ -1,4 +1,4 @@
-from .base import AnnData
+from .core.anndata import AnnData
 from .readwrite import (
     read_h5ad, read_loom, read_hdf,
     read_excel, read_umi_tools,
